requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic==2.8.0